Organize classes into packages and update auth logic
